refactor: 컴포넌트 구조 정리 및 라우팅 설정
