+++ src/components/features/Project/ProgressTracker/index.tsx
@@ -1,0 +1,3 @@
+export const ProgressTracker = () => {
+  return <div>진행도 컴포넌트</div>;
+};
