+++ src/pages/ProjectKanban/index.tsx
@@ -1,17 +1,17 @@
 import { Container, Stack } from '@chakra-ui/react';
 
-import { ProgressTracker } from '../..//components/features/Project/ProgressTracker';
+import { KanbanBoard } from '../..//components/features/Project/KanbanBoard';
 import { Project } from '../../components/features/Project';
 import { ProgressAccordion } from '../../components/features/Project/ProgressAccordion';
 
-export const ProjectPage = () => {
+export const ProjectKanbanPage = () => {
   return (
     <>
       <Project />
       <Container maxW="container.xl" padding={6}>
         <Stack spacing={6}>
           <ProgressAccordion />
-          <ProgressTracker />
+          <KanbanBoard />
         </Stack>
       </Container>
     </>
